refactor: 4단계 - JDK Dynamic Proxy를 사용해 UserService를 데이터베이스 관련 의존성과 완전히 분리하기
